requirements.txt
aniso8601==9.0.1
attrs==21.2.0
click==8.0.1
Flask==2.0.1
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
packaging==21.0
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.4
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.23
toml==0.10.2
Werkzeug==2.0.1
